pyproject.toml
[project]
name = "ha-a2a"
version = "0.1.0"
description = "Home Assistant custom integration exposing HA Assistants as A2A-compatible remote agents"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.3.0",
    "aiohttp>=3.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.11",
]

[tool.ruff]
target-version = "py313"
line-length = 88

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "UP",   # pyupgrade
    "SIM",  # flake8-simplify
    "RUF",  # ruff-specific rules
]

[tool.ruff.lint.isort]
known-first-party = ["custom_components"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra --strict-markers"
asyncio_mode = "auto"
markers = ["asyncio: mark a test as an asyncio coroutine"]

